cdk-python/requirements.txt
# aws-cdk.core>=1.132.0
# aws-cdk.assertions>=1.132.0
aws-cdk-lib==2.13.0
constructs>=10.0.0,<11.0.0
